requirements.txt
ai2thor==5.0.0
ollama
pillow
numpy
opencv-python